backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
pgvector>=0.2.4
python-dotenv>=1.0.0
google-genai>=1.0.0
docling>=2.0.0
docling-core[chunking-openai]>=2.0.0
tiktoken>=0.6.0
openai
langgraph>=0.4.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
pytest-cov>=6.0.0
pytest-html>=4.0.0
tenacity>=8.0.0
